updates from master

+++ EFCore/EFCoreSamples/MenusSample/Program.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using Microsoft.EntityFrameworkCore.ChangeTracking;
+using System;
+using System.Linq;
 
 namespace MenusSample
 {
@@ -7,21 +9,170 @@ namespace MenusSample
         static void Main()
         {
             CreateDatabase();
-            InitialFill();
+            AddRecords();
+            ObjectTracking();
+            UpdateRecords();
+            AddHundredRecords();
             DeleteDatabase();
         }
 
-        private static void InitialFill()
+        private static void AddHundredRecords()
         {
-            MenuCard[] cards = new MenuCard[]
-            {
-                new MenuCard { }
-            };
+            Console.WriteLine(nameof(AddHundredRecords));
             using (var context = new MenusContext())
             {
+                var card = context.MenuCards.FirstOrDefault();
+                if (card != null)
+                {
+                    var menus = Enumerable.Range(1, 100).Select(x => new Menu
+                    {
+                        MenuCard = card,
+                        Text = $"menu {x}",
+                        Price = 9.9m
+                    });
+                    context.Menus.AddRange(menus);
+                    int records = context.SaveChanges();
+                    Console.WriteLine($"{records} added");
+                }
+            }
+            Console.WriteLine();
+        }
 
+        private static void AddRecords()
+        {
+            Console.WriteLine(nameof(AddRecords));
+            try
+            {
+                using (var context = new MenusContext())
+                {
+
+                    var soupCard = new MenuCard();
+                    Menu[] soups =
+                    {
+                        new Menu
+                        {
+                            Text = "Consommé Célestine (with shredded pancake)",
+                            Price = 4.8m,
+                            MenuCard =soupCard
+                        },
+                        new Menu
+                        {
+                            Text = "Baked Potato Soup",
+                            Price = 4.8m,
+                            MenuCard = soupCard
+                        },
+                        new Menu
+                        {
+                            Text = "Cheddar Broccoli Soup",
+                            Price = 4.8m,
+                            MenuCard = soupCard
+                        },
+                    };
+
+                    soupCard.Title = "Soups";
+                    soupCard.Menus.AddRange(soups);
+
+                    context.MenuCards.Add(soupCard);
+
+                    ShowState(context);
+
+                    int records = context.SaveChanges();
+                    Console.WriteLine($"{records} added");
+                    Console.WriteLine();
+                }
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message);
+            }
+            Console.WriteLine();
+        }
+
+        public static void ShowState(MenusContext context)
+        {
+            foreach (EntityEntry entry in context.ChangeTracker.Entries())
+            {
+                Console.WriteLine($"type: {entry.Entity.GetType().Name}, state: {entry.State}," +
+                $" {entry.Entity}");
             }
         }
+
+        private static void ObjectTracking()
+        {
+            Console.WriteLine(nameof(ObjectTracking));
+            using (var context = new MenusContext())
+            {
+                var m1 = (from m in context.Menus
+                          where m.Text.StartsWith("Con")
+                          select m).FirstOrDefault();
+                var m2 = (from m in context.Menus
+                          where m.Text.Contains("(")
+                          select m).FirstOrDefault();
+                if (object.ReferenceEquals(m1, m2))
+                {
+                    Console.WriteLine("the same object");
+                }
+                else
+                {
+                    Console.WriteLine("not the same");
+                }
+                ShowState(context);
+            }
+            Console.WriteLine();
+        }
+
+        private static void UpdateRecords()
+        {
+            Console.WriteLine(nameof(UpdateRecords));
+            using (var context = new MenusContext())
+            {
+                Menu menu = context.Menus
+                  .Skip(1)
+                  .FirstOrDefault();
+                ShowState(context);
+                menu.Price += 0.2m;
+                ShowState(context);
+                int records = context.SaveChanges();
+                Console.WriteLine($"{records} updated");
+                ShowState(context);
+            }
+            Console.WriteLine();
+        }
+
+        private static void ChangeUntracked()
+        {
+            Menu GetMenu()
+            {
+                using (var context = new MenusContext())
+                {
+                    Menu menu = context.Menus
+                      .Skip(2)
+                      .FirstOrDefault();
+                    return menu;
+                }
+            }
+
+            Menu m = GetMenu();
+            m.Price += 0.7m;
+            UpdateUntracked(m);
+        }
+
+        private static void UpdateUntracked(Menu m)
+        {
+            using (var context = new MenusContext())
+            {
+                ShowState(context);
+                // EntityEntry<Menu> entry = context.Menus.Attach(m);
+                // entry.State = EntityState.Modified;
+                context.Menus.Update(m);
+                ShowState(context);
+                context.SaveChanges();
+            }
+        }
+
+
+
+
 
         public static void CreateDatabase()
         {
@@ -33,9 +184,14 @@ namespace MenusSample
 
         public static void DeleteDatabase()
         {
-            using (var context = new MenusContext())
+            Console.Write("Delete the database? ");
+            string input = Console.ReadLine();
+            if (input.ToLower() == "y")
             {
-                bool deleted = context.Database.EnsureDeleted();
+                using (var context = new MenusContext())
+                {
+                    bool deleted = context.Database.EnsureDeleted();
+                }
             }
         }
     }
